Outfactor generic CERL part of SBTI API
